add navigation to AddEditScheduleBlockScreen for adding

+++ app/src/main/java/com/example/schetodo/data/RoomTypeConverters.kt
@@ -17,8 +17,8 @@ class RoomTypeConverters {
     }
 
     @TypeConverter
-    fun timeStampToTime(timeStampInSeconds: Long): LocalTime {
-        return LocalTime.ofSecondOfDay(timeStampInSeconds)
+    fun timeStampToTime(timeStampInSeconds: Int): LocalTime {
+        return LocalTime.ofSecondOfDay(timeStampInSeconds.toLong())
     }
 
     @TypeConverter
